[6921] Move RemoteTabsTableViewController into separate file, per PR feedback
