+++ Client/Frontend/Browser/Tabs/RemoteTabs/RemoteTabsPanel.swift
@@ -100,218 +100,3 @@ class RemoteTabsPanel: UIViewController,
         remotePanelDelegate?.remotePanel(didSelectURL: url, visitType: VisitType.typed)
     }
 }
-
-class RemoteTabsTableViewController: UITableViewController,
-                                     Themeable,
-                                     CollapsibleTableViewSection,
-                                     LibraryPanelContextMenu {
-    struct UX {
-        static let rowHeight = SiteTableViewControllerUX.RowHeight
-    }
-
-    private(set) var state: RemoteTabsPanelState
-
-    var themeManager: ThemeManager
-    var themeObserver: NSObjectProtocol?
-    var notificationCenter: NotificationProtocol
-    weak var remoteTabsPanel: RemoteTabsPanel?
-    var tableViewDelegate: RemoteTabsPanelDataSource? {
-        didSet {
-            tableView.dataSource = tableViewDelegate
-            tableView.delegate = tableViewDelegate
-        }
-    }
-
-    private lazy var longPressRecognizer: UILongPressGestureRecognizer = {
-        return UILongPressGestureRecognizer(target: self, action: #selector(longPress))
-    }()
-
-    init(state: RemoteTabsPanelState,
-         themeManager: ThemeManager = AppContainer.shared.resolve(),
-         notificationCenter: NotificationProtocol = NotificationCenter.default) {
-        self.state = state
-        self.themeManager = themeManager
-        self.notificationCenter = notificationCenter
-        super.init(nibName: nil, bundle: nil)
-    }
-
-    required init?(coder: NSCoder) { fatalError("init(coder:) has not been implemented") }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        tableView.addGestureRecognizer(longPressRecognizer)
-        tableView.register(SiteTableViewHeader.self,
-                           forHeaderFooterViewReuseIdentifier: SiteTableViewHeader.cellIdentifier)
-        tableView.register(TwoLineImageOverlayCell.self,
-                           forCellReuseIdentifier: TwoLineImageOverlayCell.cellIdentifier)
-        tableView.register(RemoteTabsErrorCell.self,
-                           forCellReuseIdentifier: RemoteTabsErrorCell.cellIdentifier)
-
-        tableView.rowHeight = UX.rowHeight
-        tableView.separatorInset = .zero
-        tableView.alwaysBounceVertical = false
-
-        if #available(iOS 15.0, *) {
-            tableView.sectionHeaderTopPadding = 0.0
-        }
-
-        tableView.delegate = nil
-        tableView.dataSource = nil
-
-        tableView.accessibilityIdentifier = AccessibilityIdentifiers.TabTray.syncedTabs
-        listenForThemeChange(view)
-        applyTheme()
-    }
-
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        (navigationController as? ThemedNavigationController)?.applyTheme()
-
-        // Add a refresh control if the user is logged in and the control was not added before.
-        // If the user is not logged in, remove any existing control.
-        if state.allowsRefresh && refreshControl == nil {
-            addRefreshControl()
-        }
-
-        refreshTabs(state: state, updateCache: true)
-    }
-
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-        if refreshControl != nil {
-            removeRefreshControl()
-        }
-    }
-
-    func applyTheme() {
-        tableView.separatorColor = themeManager.currentTheme.colors.layerLightGrey30
-        if let delegate = tableViewDelegate as? RemoteTabsErrorDataSource {
-            delegate.applyTheme(theme: themeManager.currentTheme)
-        }
-    }
-
-    // MARK: - Refreshing TableView
-
-    func addRefreshControl() {
-        let control = UIRefreshControl()
-        control.addTarget(self, action: #selector(onRefreshPulled), for: .valueChanged)
-        refreshControl = control
-        tableView.refreshControl = control
-    }
-
-    func removeRefreshControl() {
-        tableView.refreshControl = nil
-        refreshControl = nil
-    }
-
-    @objc
-    func onRefreshPulled() {
-        refreshControl?.beginRefreshing()
-        refreshTabs(state: state, updateCache: true)
-    }
-
-    func endRefreshing() {
-        // Always end refreshing, even if we failed!
-        refreshControl?.endRefreshing()
-
-        // Remove the refresh control if the user has logged out in the meantime
-        if !state.allowsRefresh {
-            removeRefreshControl()
-        }
-    }
-
-    func updateDelegateClientAndTabData(_ clientAndTabs: [ClientAndTabs]) {
-        guard let remoteTabsPanel = remoteTabsPanel else { return }
-
-        guard !clientAndTabs.isEmpty else {
-            showEmptyTabsViewWith(.noClients)
-            return
-        }
-
-        let nonEmptyClientAndTabs = clientAndTabs.filter { !$0.tabs.isEmpty }
-        guard !nonEmptyClientAndTabs.isEmpty else {
-            showEmptyTabsViewWith(.noTabs)
-            return
-        }
-
-        let tabsPanelDataSource = RemoteTabsClientAndTabsDataSource(actionDelegate: remoteTabsPanel,
-                                                                    clientAndTabs: nonEmptyClientAndTabs,
-                                                                    theme: themeManager.currentTheme)
-        tabsPanelDataSource.collapsibleSectionDelegate = self
-        tableViewDelegate = tabsPanelDataSource
-
-        tableView.reloadData()
-    }
-
-    func refreshTabs(state: RemoteTabsPanelState, updateCache: Bool = false, completion: (() -> Void)? = nil) {
-        ensureMainThread { [self] in
-            self.state = state
-            performRefresh(updateCache: updateCache, completion: completion)
-        }
-    }
-
-    private func performRefresh(updateCache: Bool, completion: (() -> Void)?) {
-        // Short circuit if the user is not logged in
-        guard state.allowsRefresh else {
-            endRefreshing()
-            showEmptyTabsViewWith(.notLoggedIn)
-            return
-        }
-
-        // TODO: Send Redux action to get cached clients & tabs, update once new state is received. Forthcoming.
-        // store.dispatch(RemoteTabsPanelAction.refreshCachedTabs)
-    }
-
-    private func showEmptyTabsViewWith(_ error: RemoteTabsErrorDataSource.ErrorType) {
-        guard let remoteTabsPanel = remoteTabsPanel else { return }
-        var errorMessage = error
-
-        if !state.syncIsSupported { errorMessage = .syncDisabledByUser }
-
-        let remoteTabsErrorView = RemoteTabsErrorDataSource(remoteTabsDelegateProvider: remoteTabsPanel,
-                                                            error: errorMessage,
-                                                            theme: themeManager.currentTheme)
-        tableViewDelegate = remoteTabsErrorView
-
-        tableView.reloadData()
-    }
-
-    @objc
-    private func longPress(_ longPressGestureRecognizer: UILongPressGestureRecognizer) {
-        guard longPressGestureRecognizer.state == .began else { return }
-        let touchPoint = longPressGestureRecognizer.location(in: tableView)
-        guard let indexPath = tableView.indexPathForRow(at: touchPoint) else { return }
-        presentContextMenu(for: indexPath)
-    }
-
-    func hideTableViewSection(_ section: Int) {
-        guard let dataSource = tableViewDelegate as? RemoteTabsClientAndTabsDataSource else { return }
-
-        if dataSource.hiddenSections.contains(section) {
-            dataSource.hiddenSections.remove(section)
-        } else {
-            dataSource.hiddenSections.insert(section)
-        }
-
-        tableView.reloadData()
-    }
-
-    func presentContextMenu(for site: Site, with indexPath: IndexPath,
-                            completionHandler: @escaping () -> PhotonActionSheet?) {
-        guard let contextMenu = completionHandler() else { return }
-
-        present(contextMenu, animated: true, completion: nil)
-    }
-
-    func getSiteDetails(for indexPath: IndexPath) -> Site? {
-        guard let tab = (tableViewDelegate as? RemoteTabsClientAndTabsDataSource)?.tabAtIndexPath(indexPath) else {
-            return nil
-        }
-        return Site(url: String(describing: tab.URL), title: tab.title)
-    }
-
-    func getContextMenuActions(for site: Site, with indexPath: IndexPath) -> [PhotonRowActions]? {
-        return getRemoteTabContextMenuActions(for: site, remotePanelDelegate: remoteTabsPanel?.remotePanelDelegate)
-    }
-}
